"Add food recipe HTML pages
This commit introduces four new HTML pages for food recipes.
It includes the index.html page, which serves as the main landing page, and three recipe pages (butter-chicken.html, lasagna.html
", and dorayaki.html). Each page contains a header with the recipe title, an image of the dish, and lists of ingredients and steps for preparing the recipe.